improve exception handling and reorganize
