+++ lib/src/exceptions/exceptions.dart
@@ -1,6 +1,7 @@
 /// Copyright (C) 2022-2023 Intel Corporation
 /// SPDX-License-Identifier: BSD-3-Clause
 export './conditionals/conditional_exceptions.dart';
+export './interface/interface_exceptions.dart';
 export './logic/logic_exceptions.dart';
 export './logic_value/logic_value_exceptions.dart';
 export './module/module_exceptions.dart';
