+++ lib/src/exceptions/interface/interface_exceptions.dart
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 export 'interface_name_exception.dart';
+export 'interface_type_exception.dart';
