+++ lib/src/interfaces/interfaces.dart
@@ -1,0 +1,5 @@
+// Copyright (C) 2021 Intel Corporation
+// SPDX-License-Identifier: BSD-3-Clause
+
+export 'interface.dart';
+export 'pair_interface.dart';
